OI-446 base struct

See merge request dip/dip-master-data!17

+++ src/1.Core/RequestResponse/Cities/Queries/GetAll/GetAllCityQuery.cs
@@ -4,5 +4,5 @@ using Zamin.Core.RequestResponse.Queries;
 namespace Master.Data.Core.RequestResponse.Cities.Queries.GetAll;
 public sealed class GetAllCityQuery : IQuery<List<CityItemQr>>
 {
-    public int? StateId { get; set; }
+    public long? StateId { get; set; }
 }
