+++ src/2.Infra/Data/Sql.Queries/Countries/Entities/Country.cs
@@ -7,7 +7,7 @@ public sealed class Country : BaseTenantEntity
     public string DisplayTitle { get; set; } = string.Empty;
     public string CoreId { get; set; } = string.Empty;
     public string Code { get; set; } = string.Empty;
-    public int Priority { get; set; }
+    public long Priority { get; set; }
     public bool IsActive { get; set; }
     public bool IsDeleted { get; set; }
 }
